backend/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
Flask-SocketIO==5.3.6
GitPython==3.1.37
SQLAlchemy==2.0.23
python-dateutil==2.8.2
requests==2.31.0
pandas==2.1.3
numpy==1.25.2
